# Optimize encoding of 12 types still using CBOR map