Closes #764

## Description

Optimize encoding by replacing CBOR map with CBOR array for 12 Value types.

Comparisons using a 167 byte LinkValue (that doesn't contain composites and dictionaries) show:

- encoding is faster by 49.41%, alloc/op (kB) reduced by 57.82%, allocs/op reduced by 54.46%
- decoding is faster by 26.87%, alloc/op (kB) reduced by 47.60%, allocs/op unchanged

Comparisons using a 776,155 byte value with composites, dictionaries, etc. show:

- encoding is faster by 48.37%, alloc/op (kB) reduced by 47.13%, allocs/op reduced by 51.56%
- decoding is faster by 36.01%, alloc/op (kB) reduced by 31.26%, allocs/op reduced by 18.56%

<details>
  <summary> Comparisons using LinkValue :rocket:  </summary><p>

#### Encoding LinkValue

```
name                        old time/op    new time/op    delta
EncodeLink/link_167bytes-4    18.7µs ± 0%     9.5µs ± 0%  -49.41%  (p=0.000 n=10+10)
EncodeLink/link_192bytes-4    19.1µs ± 0%     9.4µs ± 0%  -50.56%  (p=0.000 n=10+10)

name                        old alloc/op   new alloc/op   delta
EncodeLink/link_167bytes-4    5.18kB ± 0%    2.18kB ± 0%  -57.82%  (p=0.000 n=10+10)
EncodeLink/link_192bytes-4    5.50kB ± 0%    2.17kB ± 0%  -60.61%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
EncodeLink/link_167bytes-4       101 ± 0%        46 ± 0%  -54.46%  (p=0.000 n=10+10)
EncodeLink/link_192bytes-4       101 ± 0%        46 ± 0%  -54.46%  (p=0.000 n=10+10)
```

#### Decoding LinkValue

```
name                        old time/op    new time/op    delta
DecodeLink/link_167bytes-4    10.6µs ± 0%     7.7µs ± 1%  -26.87%  (p=0.000 n=10+9)
DecodeLink/link_192bytes-4    10.5µs ± 0%     7.7µs ± 0%  -27.02%  (p=0.000 n=10+10)

name                        old alloc/op   new alloc/op   delta
DecodeLink/link_167bytes-4    4.50kB ± 0%    2.36kB ± 0%  -47.60%  (p=0.000 n=10+10)
DecodeLink/link_192bytes-4    4.50kB ± 0%    2.36kB ± 0%  -47.60%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
DecodeLink/link_167bytes-4      55.0 ± 0%      55.0 ± 0%     ~     (all equal)
DecodeLink/link_192bytes-4      55.0 ± 0%      55.0 ± 0%     ~     (all equal)
```
</details>

<details>
  <summary> Comparisons using composites, dictionaries, etc. :rocket:  </summary><p>

#### Encoding
```
name                                           old time/op    new time/op    delta
EncodeComposite/1comp_0dict_139bytes-4           15.9µs ± 1%     8.3µs ± 0%  -47.81%  (p=0.000 n=10+10)
EncodeComposite/2comp_0dict_160bytes-4           20.7µs ± 2%    11.7µs ± 0%  -43.62%  (p=0.000 n=10+9)
EncodeComposite/104comp_0dict_14850bytes-4        894µs ± 1%     485µs ± 1%  -45.79%  (p=0.000 n=9+10)
EncodeComposite/870comp_837dict_776155bytes-4    34.9ms ± 1%    18.0ms ± 1%  -48.37%  (p=0.000 n=10+10)

name                                           old alloc/op   new alloc/op   delta
EncodeComposite/1comp_0dict_139bytes-4           3.78kB ± 1%    1.62kB ± 0%  -57.24%  (p=0.000 n=10+10)
EncodeComposite/2comp_0dict_160bytes-4           4.87kB ± 3%    2.40kB ± 0%  -50.68%  (p=0.000 n=9+10)
EncodeComposite/104comp_0dict_14850bytes-4        247kB ± 1%     138kB ± 0%  -44.27%  (p=0.000 n=10+10)
EncodeComposite/870comp_837dict_776155bytes-4    7.46MB ± 3%    3.94MB ± 0%  -47.13%  (p=0.000 n=10+10)

name                                           old allocs/op  new allocs/op  delta
EncodeComposite/1comp_0dict_139bytes-4             76.0 ± 0%      37.0 ± 0%  -51.32%  (p=0.000 n=10+10)
EncodeComposite/2comp_0dict_160bytes-4              103 ± 0%        55 ± 0%  -46.60%  (p=0.000 n=7+10)
EncodeComposite/104comp_0dict_14850bytes-4        5.02k ± 0%     2.54k ± 0%  -49.47%  (p=0.000 n=9+10)
EncodeComposite/870comp_837dict_776155bytes-4      141k ± 0%       68k ± 0%  -51.56%  (p=0.000 n=9+10)
```

#### Decoding
```
name                                           old time/op    new time/op    delta
DecodeComposite/1comp_0dict_139bytes-4           11.5µs ± 1%     8.0µs ± 1%  -30.63%  (p=0.000 n=9+10)
DecodeComposite/2comp_0dict_160bytes-4           16.8µs ± 1%    11.6µs ± 1%  -30.90%  (p=0.000 n=10+10)
DecodeComposite/104comp_0dict_14850bytes-4        759µs ± 1%     522µs ± 1%  -31.19%  (p=0.000 n=10+10)
DecodeComposite/870comp_837dict_776155bytes-4    40.3ms ± 1%    25.8ms ± 1%  -36.01%  (p=0.000 n=10+10)

name                                           old alloc/op   new alloc/op   delta
DecodeComposite/1comp_0dict_139bytes-4           3.90kB ± 0%    2.55kB ± 0%  -34.50%  (p=0.000 n=10+10)
DecodeComposite/2comp_0dict_160bytes-4           5.54kB ± 0%    3.84kB ± 0%  -30.74%  (p=0.000 n=10+10)
DecodeComposite/104comp_0dict_14850bytes-4        296kB ± 0%     211kB ± 0%  -28.97%  (p=0.000 n=10+10)
DecodeComposite/870comp_837dict_776155bytes-4    15.0MB ± 0%    10.3MB ± 0%  -31.26%  (p=0.000 n=10+9)

name                                           old allocs/op  new allocs/op  delta
DecodeComposite/1comp_0dict_139bytes-4             56.0 ± 0%      53.0 ± 0%   -5.36%  (p=0.000 n=10+10)
DecodeComposite/2comp_0dict_160bytes-4             97.0 ± 0%      89.0 ± 0%   -8.25%  (p=0.000 n=10+10)
DecodeComposite/104comp_0dict_14850bytes-4        5.05k ± 0%     4.65k ± 0%   -7.98%  (p=0.000 n=10+10)
DecodeComposite/870comp_837dict_776155bytes-4      265k ± 0%      216k ± 0%  -18.56%  (p=0.000 n=10+10)
```

</details>

### Changes

- [x] Optimize encoding by switching from CBOR map to CBOR array for:
  - [x] StorageReferenceValue
  - [x] PathValue
  - [x] CapabilityValue
  - [x] common.AddressLocation (__see Caveats__)
  - [x] LinkValue
  - [x] CompositeStaticType
  - [x] InterfaceStaticType
  - [x] ConstantSizedStaticType
  - [x] ReferenceStaticType
  - [x] DictionaryStaticType
  - [x] RestrictedStaticType
  - [x] TypeValue (__see Caveats__)
- [x] Remove old backwards-compatibility decoding of type id in decodeCompositeStaticType() and decodeInterfaceStaticType().
- [x] Add tests, including round-trip for decoding old format and encoding new format.

### Caveats

Other CBOR representations are available and were considered for AddressLocation and TypeValue.  Simplicity and consistency were prioritized over stored data size.  My decision should be reviewed by someone more familiar with Cadence.

Benchmarks used linux_amd64 with Go 1.15.10 using production data (a snapshot of mainnet).  Other platforms, wasm, etc. may show different gains.  Potential gains from I/O reduction (from reduced stored data size) were not measured.

______

For contributor use:

- [x] Targeted PR against `feature/storage-optimizations` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels