@fxamacker 2021-04-07
> Looks great!
> 
> @fxamacker You mentioned alternatives for `AddressLocation` and `TypeValue` – what were you thinking of?

e.g. For AddressLocation, I didn't see any empty name in the v3 data of the snapshot of mainnet.  So there's opportunity to simplify by always using a 2-element array instead of having two different formats based on whether name is empty.  

**AddressLocation** is encoded as cbor tag.

 Old format's tag content is either:
  - byte slice (name is empty), or 
  - map with name and address

 New format 1 (this PR, for consistency with the old format):
  - byte slice (name is empty), or 
  - array with name and address

 New format 2 (this would simplify code, but would require 1 extra byte for empty name scenario):
  - always array with name and address

I didn't choose format 2 because it felt more like a design change vs optimization.

---

**TypeValue** is encoded as cbor tag.  

In old format, tag content is either:
  - empty map if v.Type is nil, or 
  - map of 1 pair

new format 1 (I didn't pick this because the array size isn't consistent):
  - empty array if v.Type is nil, or 
  - array of 1 element

new format 2 (this PR, array is always the same size):
  - always array of 1 element, if v.Type is nil, element is nil

new format 3 (simpler but less future proof, doesn't even use an array):
  - don't use array at all.  If v.Type is nil, tag content is nil, otherwise tag.content contains encoded v.Type

Format 2 needs less coding than format 1, but encodes 1 extra byte if v.Type is nil.
Format 3 is simpler but it would be hard to add new elements to encoded TypeValue later.