@fxamacker 2021-04-07
> @fxamacker re TypeValue #2: sounds good, as this is also consistent with all other value encodings correct? In that case it might be good to also use the array encoding for AddressLocation, the 1 byte overhead is negligible

Awesome!!!  I'll use array encoding (new format 2) for AddressLocation.  I'll whip up commit shortly.